add subtitle util
